86cym42q5: Move modal code to property-listing-page
